修复级联阴影、屏幕空间光照与 TAA，补齐烘焙光照和验证工具
解决的画面问题
- 统一级联阴影的剔除、投影和采样规则，修正阴影覆盖与级联边界异常。
- 修正 HiZ 深度金字塔的边缘和尾部 mip，统一 SSAO、SSR、SSGI 的深度重建与历史检查，避免错误命中和间接光重复叠加。
- 贯通 Lightmap、方向 Lightmap、Probe 遮蔽、Shadowmask 和 Distance Shadowmask 的数据与绘制绑定。
- 修正 Game 和 SceneView 的运动重投影及上一帧表面深度；将 TAA 累积与锐化分开，避免锐化反复写入历史造成拖尾、彩边与抖动。

实现与维护
- 按相机提交保存 Native 和 Mesh 的上一帧变换、蒙皮姿态及历史资源；恢复临时绘制属性并明确 GPU 资源释放时序。
- 补齐 Unity CLI 验证命令、正常帧采集、状态查询、取消和资源排空；修正验证期间 SceneView 关闭时机。
- 经消费者检查移除被替换的路径，补充相关回归测试、设计说明和中间数据清理规范。
- 收录本地 Unity 6.6 图标导入元数据更新，移除已合并进 Collections 的旧 Jobs 依赖，并修正64位对象标识转换。

已记录的验证
- 第一轮视觉修复 Editor 测试241/241通过；最终 TAA 候选 Editor 测试264/264通过。
- 检查了 Game/SceneView 正常运动帧、原始 GPU 数据、1倍/10倍窗口和停止运动后的放大序列。
- 最终候选两条30分钟空闲观察未复现崩溃；独立 Astra low 核验通过渲染范围，但未通过完整崩溃归因门。
- 以上为本轮此前完成的测试记录，本次提交整理未重新运行全部测试。

仍未解决的事项
- Unity6000.6.0f1原生UDS误解析ExFAT上的AppleDouble伴随文件，原路径启动仅临时恢复，新文件仍会触发复发；未宣称持久修复。
- 更早的空闲SourceAssetDB/MDB_BAD_RSLOT崩溃仍未定因，未认定为InfinityRP或引擎单方导致。
- 自定义Mesh透明绘制的既有N11缺口与其他未测平台不在本次通过范围。

交付边界
- 按用户要求提交当前本地change list，不代表两份计划已经全部验收完成。
- 未迁移项目、修改或重签名Unity安装程序、升级引擎、重烘焙资产或构建macOS Player。
- 仅提交本包内文件；Example内嵌CLI桥接依赖和本地诊断原始数据不包含在此仓库。

+++ Runtime/Component/Render/LightComponent.cs
@@ -12,7 +12,7 @@ namespace InfinityTech.Component
     public class LightComponent : BaseComponent
     {
         Light m_UnityLight;
-        int m_RegisteredLightId;
+        ulong m_RegisteredLightId;
         public Light unityLight => m_UnityLight ? m_UnityLight : (m_UnityLight = GetComponent<Light>());
         public ERenderingLayer lightLayer = ERenderingLayer.LightLayerDefault;
 
@@ -37,8 +37,8 @@ namespace InfinityTech.Component
         protected override void OnRegister()
         {
             m_UnityLight = GetComponent<Light>();
-            m_RegisteredLightId = UnityEntityId.ToInt32(m_UnityLight);
-            int lightId = m_RegisteredLightId;
+            m_RegisteredLightId = UnityEntityId.ToUInt64(m_UnityLight);
+            ulong lightId = m_RegisteredLightId;
             FGraphics.AddTask((RenderContext renderContext) =>
             {
                 if (this && m_UnityLight) renderContext.AddWorldLight(lightId, this);
@@ -57,7 +57,7 @@ namespace InfinityTech.Component
 
         protected override void UnRegister()
         {
-            int lightId = m_RegisteredLightId;
+            ulong lightId = m_RegisteredLightId;
             FGraphics.AddTask((RenderContext renderContext) =>
             {
                 renderContext.RemoveWorldLight(lightId);
